+++ js-performance.js
@@ -10,7 +10,6 @@ const appRoot = __dirname;
 global.requireModule = (module) => require(join(appRoot, module));
 const DEFAULTS = requireModule('src/support/defaults');
 const VERBOSITY = requireModule('src/support/verbosity');
-const testdata = requireModule('src/support/testdata/testdata');
 const ProfileRunner = requireModule('src/profile-runner');
 const Reporter = requireModule('src/reporter/reporter');
 
@@ -18,8 +17,9 @@ const opts = new GetOpt([
     ['h', 'help', 'Display this helptext.'],
     ['i', 'iterations=', `Specify the number of iterations per profiled function. Default: ${DEFAULTS.iterations}.`],
     ['l', 'list', 'List available profiles.'],
-    ['q', 'quiet', 'Print results only.'],
     ['m', 'magnitude=', `Specify the magnitude of testdata. Default: ${DEFAULTS.testdataMagnitude}.`],
+    ['p', 'precision=', `Specify the precision in terms of decimal places of results. Default: ${DEFAULTS.precision} decimals.`],
+    ['q', 'quiet', 'Print results only.'],
     ['v', 'verbose', 'Print verbose information.']
   ]).bindHelp()
   .parseSystem();
@@ -47,12 +47,16 @@ if ('verbose' in opts.options) {
   verbosity = VERBOSITY.VERBOSE;
 }
 
-let data;
+let magnitude = DEFAULTS.testdataMagnitude;
 if ('magnitude' in opts.options
   && !isNaN(parseInt(opts.options.magnitude, 10))) {
-  data = testdata(parseInt(opts.options.magnitude, 10));
-} else {
-  data = testdata();
+  magnitude = parseInt(opts.options.magnitude, 10);
+}
+
+let precision = DEFAULTS.precision;
+if ('precision' in opts.options
+  && !isNaN(parseInt(opts.options.precision, 10))) {
+  precision = parseInt(opts.options.precision, 10);
 }
 
 let profiles = [];
@@ -72,8 +76,8 @@ if (opts.argv.length > 0) {
 const profileRunner = new ProfileRunner({
   iterations,
   profiles,
-  data
+  magnitude
 });
 
-new Reporter(profileRunner, verbosity, DEFAULTS.timeout);
+new Reporter(profileRunner, verbosity, precision);
 profileRunner.run();