+++ src/profile-runner.js
@@ -3,6 +3,7 @@
 const EventEmitter = require('events');
 const clock = requireModule('src/support/clock/clock');
 const events = requireModule('src/support/events');
+const testdata = requireModule('src/support/testdata/testdata');
 
 class ProfileRunner extends EventEmitter {
 
@@ -11,13 +12,13 @@ class ProfileRunner extends EventEmitter {
   * @param {object} config Configuration object
   * @param {array} config.profiles An array of profiles
   * @param {number} config.iterations Number of iterations
-  * @param {array} config.data An array of test data
+  * @param {number} config.magnitude Magnitude of test data
   */
   constructor(config) {
     super();
     this.profiles = config.profiles.slice();
     this.iterations = config.iterations;
-    this.data = config.data;
+    this.magnitude = config.magnitude;
   }
 
   /**
@@ -44,7 +45,7 @@ class ProfileRunner extends EventEmitter {
     const result = {
       profile,
       testResults: profile.functions.map((func) =>
-        this.runFunction(profile, func))
+        this.runFunction(profile, func, testdata(profile.testDataType, this.magnitude)))
     };
     this.emit(events.PROFILE_END, profile, result);
     return result;
@@ -55,9 +56,10 @@ class ProfileRunner extends EventEmitter {
   * @private
   * @param {object} profile A profile
   * @param {object} func A profile function
+  * @param {any} data The test data
   * @returns {object} The test result
   */
-  runFunction(profile, func) {
+  runFunction(profile, func, data) {
     this.emit(events.TEST_START, profile, func);
     const testResult = {
       averageTime: 0,
@@ -66,7 +68,7 @@ class ProfileRunner extends EventEmitter {
     };
 
     for (let i = 0; i < this.iterations; i++) {
-      testResult.totalTime += clock.time(func.f, this.data);
+      testResult.totalTime += clock.time(func.f, data);
     }
 
     testResult.averageTime = testResult.totalTime / this.iterations;