+++ src/profiles/loops/loops.profile.js
@@ -113,5 +113,6 @@ module.exports = {
     loopMap,
     loopWhile,
     loopWhileFixedLen
-  ]
+  ],
+  testDataType: 'array'
 };