+++ src/reporter/reporter.js
@@ -8,9 +8,11 @@ class Reporter {
   /**
   * @param {ProfileRunner} profileRunner The test runner
   * @param {number} verbosity The verbosity level
+  * @param {number} precision The number of decimal places to print for results
   */
-  constructor(profileRunner, verbosity) {
+  constructor(profileRunner, verbosity, precision) {
     this.verbosity = verbosity;
+    this.precision = precision;
 
     profileRunner.on(events.START, (profiles) => {
       if (this.verbosity === VERBOSITY.VERBOSE) {
@@ -31,9 +33,17 @@ class Reporter {
     });
 
     profileRunner.on(events.PROFILE_END, (profile, result) => {
-      const best = result.testResults
-        .sort((a, b) => a.averageTime - b.averageTime)[0];
-      console.info(chalk.green(`  Fastest: "${best.func.description()}" (${best.averageTime.toFixed(3)}ms)\n`));
+      if (this.verbosity >= VERBOSITY.NORMAL) {
+        console.log('  ==============');
+        const best = result.testResults
+          .sort((a, b) => a.averageTime - b.averageTime)
+          .filter((a, i, arr) => a.averageTime === arr[0].averageTime);
+        if (best.length === 1) {
+          console.info(chalk.green(`  Winner: "${best[0].func.description()}" (${this.formatTime(best[0].averageTime)})\n`));
+        } else {
+          console.info(chalk.green(`  Draw between:\n${best.forEach((b) => `\t"${b.func.description()}" (${this.formatTime(b.averageTime)})`)}`));
+        }
+      }
     });
 
     profileRunner.on(events.TEST_START, (profile, func) => {
@@ -44,14 +54,21 @@ class Reporter {
 
     profileRunner.on(events.TEST_END, (profile, func, result) => {
       if (this.verbosity >= VERBOSITY.NORMAL) {
-        console.info(`    ${result.averageTime.toFixed(3)}ms`);
-        console.info();
+        console.info(`    ${this.formatTime(result.averageTime)}\n`);
       }
     });
 
     profileRunner.on(events.ERROR, (err) => {
       console.error(err.stack);
     });
+  }
+
+  formatTime(t) {
+    if (t * 1000 < 10) {
+      return `${(t * 1000).toFixed(this.precision)}μs`;
+    }
+
+    return `${t.toFixed(this.precision)}ms`;
   }
 }