+++ src/support/defaults.js
@@ -5,5 +5,6 @@ const VERBOSITY = requireModule('src/support/verbosity');
 module.exports = {
   iterations: 1000,
   verbosity: VERBOSITY.NORMAL,
-  testdataMagnitude: 1000
+  testdataMagnitude: 1000,
+  precision: 4
 };